Filter triples that need to be sent
Before this commit, if one of the triples in the group of changeSets
matches a specification, the whole group is sent to the specified host.
This means that whenever a lot of random data enters the delta-notifier
simultaneously, but one of those triples happens to match a specification,
the whole set of data is sent to that host even if that host is only
interested in that single triple. Think about the implication with
relation to sensitive data!!!
This commit filters the changeSets to only include on a per-triple level
what is actually needed for that host, while still keepin the changeSets
in order. This means that the host only receives the triples it is
interested in / allowed to see.