Experiment to propagate allowed-groups to clients
We probably want to have some structure to better filter out quads and better
filter out information sent to consumers (also see #26).

This experiment allows dropping the access rights used in the request's header.

It doesn't allow dropping the mu-session-id though doing so could allow for more
involved folding and we see no reason not allow this too.

The mu-auth-allowed-groups are still persisted in the delta body.  It's not
documented to be available there so we could remove it there too.

A function can be specified to alter the access rights.